implement the exucition of invastigate, fixed few crash bugs, and tested the system, need a quit fix of numbers. but the idea is working, the numbers ####. also find few logic holes (ft is can change the pointB and that can change to where walk the npc.) and the numbers needed to be changes, the scary grow too fast.

+++ Source/K7/DmgDealer.h
@@ -15,7 +15,7 @@ public:
 	// Sets default values for this actor's properties
 	ADmgDealer();
 	UPROPERTY(EditAnywhere,BlueprintReadOnly)
-	int32 IDcur;
+	uint8 IDcur;
 	UPROPERTY(EditAnywhere,BlueprintReadOnly)
 	TArray<AActor*> regOb;
 	UPROPERTY(EditAnywhere,BlueprintReadOnly)
